Move all GraphQL types to src/server/types
