+++ .vscode/snippets/javascript.json
@@ -1,4 +1,36 @@
 {
+  "Route": {
+    "prefix": "route",
+    "body": [
+      "import React from 'react';",
+      "import { graphql } from 'relay-runtime';",
+      "import Layout from '../common/Layout';",
+      "",
+      "export default [",
+      "  {",
+      "    path: '/${1:path}',",
+      "    query: graphql`",
+      "      query ${TM_DIRECTORY/.*[\\/](.*)$/$1/}${2:Page}Query {",
+      "        ...Layout_data",
+      "        ...${2:Page}_data",
+      "      }",
+      "    `,",
+      "    components: () => [import(/* webpackChunkName: '${1:path}' */ './${2:Page}')],",
+      "    render: ([${2:Page}], data, { config }) => ({",
+      "      title: `${3:Title} • ${config.app.name}`,",
+      "      component: (",
+      "        <Layout data={data}>",
+      "          <${2:Page} data={data} />",
+      "        </Layout>",
+      "      ),",
+      "      chunks: ['${1:path}'],",
+      "    }),",
+      "  },",
+      "];",
+      ""
+    ],
+    "description": "Route"
+  },
   "ReactComponent": {
     "prefix": "reactComponent",
     "body": [
@@ -76,16 +108,13 @@
       "  );",
       "}",
       "",
-      "export default createFragmentContainer(",
-      "  ${1:Component},",
-      "  {",
-      "    data: graphql`",
-      "      fragment ${1:Component}_data on Query {",
-      "        id",
-      "      }",
-      "    `,",
-      "  },",
-      ");",
+      "export default createFragmentContainer(${1:Component}, {",
+      "  data: graphql`",
+      "    fragment ${1:Component}_data on Query {",
+      "      id",
+      "    }",
+      "  `,",
+      "});",
       ""
     ],
     "description": "React/Relay Fragment Container"
