+++ src/index.js
@@ -10,9 +10,9 @@ import qs from 'query-string';
 import { createBrowserHistory } from 'history';
 
 import App from './common/App';
-import createRelay from './createRelay';
 import * as serviceWorker from './serviceWorker';
 import router from './router';
+import { createRelay } from './relay';
 import { setHistory } from './utils/scrolling';
 
 const container = document.getElementById('root');
