+++ src/server/mutations/index.js
@@ -4,7 +4,5 @@
  * Copyright (c) 2015-present Kriasoft | MIT License
  */
 
-export * from './upsertStory';
-export * from './likeStory';
-export * from './updateUser';
-export * from './deleteUser';
+export * from './user';
+export * from './story';
