+++ src/server/queries/index.js
@@ -4,8 +4,5 @@
  * Copyright (c) 2015-present Kriasoft | MIT License
  */
 
-export { default as me } from './me';
-export { default as user } from './user';
-export { default as users } from './users';
-export { default as story } from './story';
-export { default as stories } from './stories';
+export * from './user';
+export * from './story';
