+++ src/server/queries/user.js
@@ -4,11 +4,26 @@
  * Copyright (c) 2015-present Kriasoft | MIT License
  */
 
-import { GraphQLNonNull, GraphQLString } from 'graphql';
+import { GraphQLNonNull, GraphQLInt, GraphQLString } from 'graphql';
+import {
+  connectionDefinitions,
+  forwardConnectionArgs,
+  connectionFromArraySlice,
+  cursorToOffset,
+} from 'graphql-relay';
 
-import { UserType } from '../user';
+import db from '../db';
+import { UserType } from '../types';
 
-export default {
+export const me = {
+  type: UserType,
+
+  resolve(root, args, ctx) {
+    return ctx.user ? ctx.userById.load(ctx.user.id) : null;
+  },
+};
+
+export const user = {
   type: UserType,
 
   args: {
@@ -17,5 +32,49 @@ export default {
 
   resolve(root, { username }, ctx) {
     return ctx.userByUsername.load(username);
+  },
+};
+
+export const users = {
+  type: connectionDefinitions({
+    name: 'User',
+    nodeType: UserType,
+    connectionFields: {
+      totalCount: { type: new GraphQLNonNull(GraphQLInt) },
+    },
+  }).connectionType,
+
+  args: forwardConnectionArgs,
+
+  async resolve(root, args, ctx) {
+    // Only admins allowed to fetch the list of users
+    ctx.ensureIsAuthorized(user => user.isAdmin);
+
+    const limit = typeof args.first === 'undefined' ? '100' : args.first;
+    const offset = args.after ? cursorToOffset(args.after) + 1 : 0;
+
+    const [data, totalCount] = await Promise.all([
+      db
+        .table('users')
+        .orderBy('created_at', 'desc')
+        .limit(limit)
+        .offset(offset)
+        .then(rows => {
+          rows.forEach(x => ctx.userById.prime(x.id, x));
+          return rows;
+        }),
+      db
+        .table('users')
+        .count()
+        .then(x => x[0].count),
+    ]);
+
+    return {
+      ...connectionFromArraySlice(data, args, {
+        sliceStart: offset,
+        arrayLength: totalCount,
+      }),
+      totalCount,
+    };
   },
 };
