--- src/server/queries/users.js
@@ -1,60 +1,0 @@
-/**
- * React Starter Kit for Firebase
- * https://github.com/kriasoft/react-firebase-starter
- * Copyright (c) 2015-present Kriasoft | MIT License
- */
-
-import { GraphQLNonNull, GraphQLInt } from 'graphql';
-import {
-  connectionDefinitions,
-  forwardConnectionArgs,
-  connectionFromArraySlice,
-  cursorToOffset,
-} from 'graphql-relay';
-
-import db from '../db';
-import { UserType } from '../user';
-
-export default {
-  type: connectionDefinitions({
-    name: 'User',
-    nodeType: UserType,
-    connectionFields: {
-      totalCount: { type: new GraphQLNonNull(GraphQLInt) },
-    },
-  }).connectionType,
-
-  args: forwardConnectionArgs,
-
-  async resolve(root, args, ctx) {
-    // Only admins allowed to fetch the list of users
-    ctx.ensureIsAuthorized(user => user.isAdmin);
-
-    const limit = typeof args.first === 'undefined' ? '100' : args.first;
-    const offset = args.after ? cursorToOffset(args.after) + 1 : 0;
-
-    const [data, totalCount] = await Promise.all([
-      db
-        .table('users')
-        .orderBy('created_at', 'desc')
-        .limit(limit)
-        .offset(offset)
-        .then(rows => {
-          rows.forEach(x => ctx.userById.prime(x.id, x));
-          return rows;
-        }),
-      db
-        .table('users')
-        .count()
-        .then(x => x[0].count),
-    ]);
-
-    return {
-      ...connectionFromArraySlice(data, args, {
-        sliceStart: offset,
-        arrayLength: totalCount,
-      }),
-      totalCount,
-    };
-  },
-};
