--- src/server/story/index.js
@@ -1,8 +1,0 @@
-/**
- * React Starter Kit for Firebase
- * https://github.com/kriasoft/react-firebase-starter
- * Copyright (c) 2015-present Kriasoft | MIT License
- */
-
-export { default as StoryType } from './StoryType';
-export { default as CommentType } from './CommentType';
